fixed a bug

+++ team8/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class LearningWord(models.Model):
-    user_id = models.IntegerField() 
+    user_id = models.BigIntegerField() 
     word = models.CharField(max_length=100)
     
     ipa_pronunciation = models.CharField(max_length=100, blank=True)
